правки для gcc.

+++ xray/Layers/xrRender/SkeletonAnimated.cpp
@@ -82,7 +82,7 @@ static LPCSTR name_blend_type( CBlend::ECurvature blend )
 		{"eFREE_SLOT"		, CBlend::eFREE_SLOT	},
 		{"eAccrue"			, CBlend::eAccrue		},
 		{"eFalloff"			, CBlend::eFalloff		},
-		{"eFORCEDWORD"		, CBlend::eFORCEDWORD	}
+		{"eFORCEDWORD"		, (int)CBlend::eFORCEDWORD	}
 	};
 	return get_token_name( token_blend, blend );
 }
